pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "evolutionary-snake"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "coverage (>=7.8.2,<8.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "neat-python (>=0.92,<0.93)",
]

[tool.poetry]
packages = [{include = "evolutionary_snake", from = "src"}]
requires-poetry = ">=2.1"

[tool.poetry.scripts]
snake = "evolutionary_snake.cli:start"
training = "evolutionary_snake.cli:start_training"

[tool.poetry.group.dev.dependencies]
tox = "^4.25.0"
ruff = "^0.11.6"
pylint = "^3.3.6"
bandit = "^1.8.3"
mypy = "^1.15.0"
pytest = "^8.3.5"

# https://coverage.readthedocs.io/en/7.2.2/config.html
[tool.coverage.run]
branch = true
parallel = true
source = ["evolutionary_snake"]
omit = ["cli.py"]

[tool.coverage.report]
fail_under = 100
precision = 0
show_missing = true
skip_covered = false
skip_empty = false
sort = "Cover"
exclude_also = [
    "if TYPE_CHECKING:"
]

[tool.coverage.paths]
source = [
    "src",
    ".tox/*/lib/site-packages",
    ".tox/*/lib/python*/site-packages",
    ".tox/pypy*/site-packages",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage-report.xml"

[tool.pylint]
disable = [
    "W0511", # https://pylint.pycqa.org/en/latest/user_guide/messages/warning/fixme.html
    "C0411", # Conflicts with ruff's isort
    "W0613", # unused-argument, this is covered by Ruff ARG001
    "W0611", # unused-import, covered by Ruff F401
    "C0301", # line-too-long, covered by Ruff E501
    "W0718", # broad-exception-caught, covered by Ruff BLE001
    "W0719", # broad-exception-raised, covered by Ruff TRY002
    "W0212", # private member accessed, covered by Ruff SLF001
    "R0913", # too-many-arguments, covered by Ruff PLR0913
]
extension-pkg-allow-list = ["pygame"]

# https://docs.astral.sh/ruff/configuration/
[tool.ruff]
extend = "ruff.toml"
src = ["src", "tests"]

# https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml
[tool.mypy]
strict = true
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
plugins = ["pydantic.mypy"]
